+++ CHANGELOG.md
@@ -9,12 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- use of OTP-trusted CAs on OTP 25+
 - Windows to CI
 
 ### Changed
 
-- CAs to the ones trusted by OTP (typically provided by the OS), when available, on OTP 25+
+- default CAs to the ones trusted by OTP (typically provided by the OS), when available, on OTP 25+
 
 ## [1.16.0] - 2022-10-11